--- Bogcha.Domain/Entities/Models/Attendence.cs
@@ -1,6 +1,0 @@
-﻿namespace Bogcha.Domain.Entities.Models
-{
-    public class Attendence
-    {
-    }
-}
