--- Bogcha.Domain/Entities/Models/Student.cs
@@ -1,6 +1,0 @@
-﻿namespace Bogcha.Domain.Entities.Models
-{
-    public class Student
-    {
-    }
-}
