--- Bogcha.Domain/Entities/Models/Teacher.cs
@@ -1,6 +1,0 @@
-﻿namespace Bogcha.Domain.Entities.Models
-{
-    public class Teacher
-    {
-    }
-}
