+++ Bogcha.Domain/Entities/Teacher.cs
@@ -1,0 +1,6 @@
+﻿namespace Bogcha.Domain.Entities
+{
+    public class Teacher
+    {
+    }
+}
