+++ Bogcha.Domain/Enums/Attendence_Status_Enum.cs
@@ -1,0 +1,8 @@
+﻿namespace Bogcha.Domain.Enums
+{
+    public enum Attendence_Status_Enum
+    {
+        False,
+        True
+    }
+}
